fist commit with the basic scripts to create an OctTree
This first commit is able to succesfully create and plot the OctTree structrure for a given number of particles. It has to be fixed that the plotting method does not update to the size of the region in where the particles are, since it only plots between 0 and 1 for all the axis. This is anyways a minor problem and will be solved by setting the initial limits